test: seed numpy RNG for snapshot determinism + move build_model_grid to _constants
